+++ src/Security/Voter/OrderVoter.php
@@ -7,6 +7,7 @@ use App\Entity\Order;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
+use Symfony\Component\Security\Core\Authorization\Voter\Vote;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
@@ -36,7 +37,10 @@ final class OrderVoter extends Voter
             && $subject instanceof \App\Entity\Order;
     }
 
-    protected function voteOnAttribute(string $attribute, mixed $subject, TokenInterface $token): bool
+    protected function voteOnAttribute(string $attribute, 
+                                        mixed $subject, 
+                                        TokenInterface $token,
+                                        ?Vote $vote = null): bool
     {
         $user = $token->getUser();