+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Tous les changements notables de ce projet seront documentés dans ce fichier.
 ### Added
 - Possibilité de définir une adresse principale directement depuis la page "Mes adresses"
 
+### Updated
+- Mise a jour du framework Symfony 7.2 vers 7.3
+
 ## [1.0.1] - 2025-09-09
 ### Added
 - Première version du projet avec système d'authentification client et gestion du compte.